+++ foliant/cli/make.py
@@ -88,8 +88,14 @@ class Cli(BaseCli):
             raise BackendError('No backend specified')
 
     @ignore
-    def get_config(self, project_path: Path, config_file_name: str, debug=False) -> dict:
-        with spinner('Parsing config', self.logger, debug):
+    def get_config(
+            self,
+            project_path: Path,
+            config_file_name: str,
+            quiet=False,
+            debug=False
+    ) -> dict:
+        with spinner('Parsing config', self.logger, quiet, debug):
             try:
                 config = Parser(project_path, config_file_name, self.logger).parse()
 
@@ -112,8 +118,9 @@ class Cli(BaseCli):
         {
             'target': 'Target format: pdf, docx, html, etc.',
             'backend': 'Backend to make the target with: Pandoc, MkDocs, etc.',
-            'project_path': 'Path to the Foliant project',
-            'config_file_name': 'Name of config file of the Foliant project',
+            'project_path': 'Path to the Foliant project.',
+            'config_file_name': 'Name of config file of the Foliant project.',
+            'quiet': 'Hide all output accept for the result. Useful for piping.',
             'keep_tmp': 'Keep the tmp directory after the build.',
             'debug': 'Log all events during build. If not set, only warnings and errors are logged.'
         }
@@ -124,6 +131,7 @@ class Cli(BaseCli):
             backend='',
             project_path=Path('.'),
             config_file_name='foliant.yml',
+            quiet=False,
             keep_tmp=False,
             debug=False
         ):
@@ -143,7 +151,7 @@ class Cli(BaseCli):
             else:
                 backend = self.get_matching_backend(target, available_backends)
 
-            config = self.get_config(project_path, config_file_name, debug)
+            config = self.get_config(project_path, config_file_name, quiet, debug)
 
         except (BackendError, ConfigError) as exception:
             self.logger.critical(str(exception))
@@ -163,14 +171,18 @@ class Cli(BaseCli):
             result = backend_module.Backend(
                 context,
                 self.logger,
+                quiet,
                 debug
             ).preprocess_and_make(target)
 
         if result:
             self.logger.info(f'Result: {result}')
 
-            print('─' * 20)
-            print(f'Result: {result}')
+            if not quiet:
+                print('─' * 20)
+                print(f'Result: {result}')
+            else:
+                print(result)
 
         else:
             self.logger.critical('No result returned by backend')